+++ src-docs/src/views/popover/popover_panel_class_name.js
@@ -26,7 +26,6 @@ export default class extends Component {
   render() {
     return (
       <EuiPopover
-        id="popoverPanelClassName"
         ownFocus
         button={
           <EuiButton